# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dummy Backend API",
    "description": "API documentation for the Dummy Backend system - managing HR, Payroll, Operations, Inventory, and Fleet Management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
...
  "paths": {
...
    "/clean/op_bus-trip-details": {
      "get": {
        "tags": ["Bus Operations"],
        "summary": "Get bus trip details",
        "description": "Returns bus trip details with driver/conductor information, expenses, and revenue",
        "parameters": [
          {
            "name": "RequestType",
            "in": "query",
            "description": "Filter by request type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["revenue", "expense"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BusTripDetail" }
                },
                "example": [
                  {
                    "assignment_id": "BA-yfu9rzboxjfob4rl8gn0psfp",
                    "bus_trip_id": "BT-n7w2yza9r2l6t1q6w31txyea",
                    "bus_route": "Sapang Palay - PITX",
                    "is_revenue_recorded": false,
                    "is_expense_recorded": false,
                    "date_assigned": "2026-01-10T00:00:00.000Z",
                    "trip_fuel_expense": 200,
                    "trip_revenue": 1000,
                    "assignment_type": "PERCENTAGE",
                    "assignment_value": 0.1,
                    "payment_method": "Reimbursement",
                    "employee_driver": {
                      "employeeId": "clxyz123",
                      "employeeNumber": "EMP-2021-OPS-D001",
                      "firstName": "Mario",
                      "lastName": "Perez"
                    },
                    "employee_conductor": {
                      "employeeId": "clxyz456",
                      "employeeNumber": "EMP-2021-OPS-C001",
                      "firstName": "Jose",
                      "lastName": "Domingo"
                    },
                    "bus_id": "BUS001",
                    "bus_plate_number": "PLATE-0001",
                    "bus_type": "Aircon",
                    "body_number": "BODY-0001"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": ["Bus Operations"],
        "summary": "Update bus trip recorded flags",
        "description": "Update IsRevenueRecorded and IsExpenseRecorded flags for bus trips",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["bus_trip_id"],
                  "properties": {
                    "bus_trip_id": { "type": "string" },
                    "IsRevenueRecorded": { "type": "boolean" },
                    "IsExpenseRecorded": { "type": "boolean" }
                  }
                }
              },
              "example": [
                {
                  "bus_trip_id": "BT-n7w2yza9r2l6t1q6w31txyea",
                  "IsRevenueRecorded": true,
                  "IsExpenseRecorded": true
                }
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": { "type": "array" },
                    "failed": { "type": "array" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Employee": {
        "type": "object",
        "properties": {
          "employeeNumber": { "type": "string", "example": "EMP-2021-OPS-007" },
          "firstName": { "type": "string", "example": "Mario" },
          "middleName": { "type": "string", "example": "Ramos" },
          "lastName": { "type": "string", "example": "Perez" },
          "phone": { "type": "string", "example": "+639389025401" },
          "position": { "type": "string", "example": "Driver" },
          "barangay": { "type": "string", "example": "Santo Niño" },
          "zipCode": { "type": "string", "example": "1100" },
          "departmentId": { "type": "integer", "example": 4 },
          "department": { "type": "string", "example": "Operations" }
        }
      },
...
      "BusTripDetail": {
        "type": "object",
        "properties": {
          "assignment_id": { "type": "string" },
          "bus_trip_id": { "type": "string" },
          "bus_route": { "type": "string" },
          "is_revenue_recorded": { "type": "boolean" },
          "is_expense_recorded": { "type": "boolean" },
          "date_assigned": { "type": "string", "format": "date-time" },
          "trip_fuel_expense": { "type": "number" },
          "trip_revenue": { "type": "number" },
          "assignment_type": { "type": "string", "enum": ["FIXED", "PERCENTAGE"] },
          "assignment_value": { "type": "number" },
          "payment_method": { "type": "string", "enum": ["Cash", "Reimbursement", "RenterDamage"] },
          "employee_driver": { "$ref": "#/components/schemas/Employee" },
          "employee_conductor": { "$ref": "#/components/schemas/Employee" },
          "bus_id": { "type": "string" },
          "bus_plate_number": { "type": "string" },
          "bus_type": { "type": "string" },
          "body_number": { "type": "string" }
        }
      },
...
    }
  }
}